pyproject.toml
[tool.poetry]
name = "kaizen"
version = "0.1.7"
description = "A tool for continuous improvement"
authors = ["Thanassis Avgerinos <thanassis.avgerinos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.43"
pygithub = "^2.5.0"
neurosymbolic = "0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kaizen = "kaizen:main"
